Chart from library didn't fit, switching to another

+++ app/src/main/java/com/pickyberry/rtuitlab_recruit/data/network/details/CoinDetailsMarketCapDto.kt
@@ -3,6 +3,6 @@ package com.pickyberry.rtuitlab_recruit.data.network.details
 import com.google.gson.annotations.SerializedName
 
 data class CoinDetailsMarketCapDto(
-    @SerializedName("rub") var rub: Int? = null,
-    @SerializedName("usd") var usd: Int? = null,
+    @SerializedName("rub") var rub: Double? = null,
+    @SerializedName("usd") var usd: Double? = null,
 )
